Logging Mechanism Added to the code
