+++ .gitignore
@@ -13,6 +13,7 @@ user_ip_wavfile_folder
 Slakh2100
 Outputs
 DataBase
+Logs
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
